☀️ 20% – Back out 30 files of evidence-free retype edits
`1c44782f6` kept whatever a greedy sweep over seven return types happened to
move, on owners hundreds of halfwords from matching, and said so in its own
message: "No new bytes." That is not a cheap improvement, it is a false claim
written into the corpus. One of its hunks flips four callees from `void` to
`s32` because the positional score dropped by two; nothing establishes that any
of them returns anything.

Positional halfword counts are only comparable between candidates of identical
size, and are close to meaningless at that distance -- a rule this repository's
own ledger states and which the commit quoted before ignoring.

30 of the 34 files are restored. Three are left because the pool-word commit
later edited them on real evidence, and one is now exact.

What stays is what a fact produced: the return-type rule where it CLOSED an
owner, which is verified by bytes, and the pool-word rule, which encodes a
stated law rather than a search. The build still reproduces the ROM
byte-identically.

+++ semantic/resource_3bb_c_02002e84.c
@@ -21,7 +21,7 @@ extern u16 Data_0200cc5c;
 extern s16 Data_0200cc14;
 extern s16 Data_0200cbf8;
 
-extern s32 Func_03000380(s32 value, u8 divisor);
+extern s32 Func_03000380(s32 value, s32 divisor);
 extern void Func_080001e8(void *command, s32 opacity);
 
 void Func_02002e84(void)
